Get currently playing track of authorized user
Clicking button in index.html now executes oauth2.js code.

oauth2.js prompts the user to log in to authorize Adepto to gain access
to their account details specifically, what track they are currently
playing. This is followed by using the authorization code retrieved
from logging in to send a POST request to get refresh and access tokens.
The access token is then used in a GET request to obtain the track the user is
currently playing. The name of the track is then printed to the console.

There is no error handling yet in instances where responses have no data
populated or where there is a 204 NO CONTENT response.