Separate each authorization step
Each step of the authorization process is separated and can be invoked
by clicking on their respective button added to index.html. This is in
preparation for when the buttons are located in different places in the
future.

Some buttons must be clicked on before other buttons because some code
is dependent on values received from other functions. Some values can
only be used once.