Check that sent state is same as response state
A query parameter called state is sent with the first authorization
request. It is used to help match requests and responses.

Since the redirect_uri query parameter can be guessed, state can be used
to make sure that the incoming connection is the result of an
authentication request.

A function has been added to generate a random string on every
authorization request.

The sent state is checked to see if it matches with the response state.
For now, the result of this check is printed to the console.