Set and get access and refresh tokens as cookie
This requires adding the cookies permission in the manifest.json.

Cookies can't be accessed through the console because I believe that
the cookies are not set in index.html. This is based on this answer:
https://stackoverflow.com/a/13541843

They are still set within the browser. They can be accessed through Site
Settings > Cookies > See all cookies and site data.