Request for refreshed access token when expiring
The access token provided by Spotify is deliberately set to expire after
one hour, so an expiration date is added to the access token cookie when it is
being set. This is so that prior to requesting for data about the currently
playing track, the current time can be compared to the set expiration date. If
the current time is at least a minute before the expiry, it will still
be used to request for data. Otherwise, a refreshed access token will be
requested and set as a cookie. A request for data is attempted again
automatically with the refreshed access token without clicking the button
again.

For now, I check to see that the current time is at least a minute
before the expiry because it provides extra time for requesting with
the access token that will expire soon.