Move request for current track to background.js
The plan is that index.html might move to a popup instead where the main
purpose is just to log in. This means that index.html no longer is able
to request for data. Requesting data for the currently playing
track will still be done through the browserAction.

There is kind of a bug after logging in. It is not possible to request
data right away. There seems to be a delay when setting the access and
refresh tokens as a cookie and getting them. When I request for data,
the access token will not exist for a few seconds. When it does exist,
it will still fail to request for data.