Get cookies with one function call
chrome.cookies.getAll() is used instead consolidate token getting. This is
so that the requesting data or requesting a new access token can be done
within the same callback function block.

Setting the refresh token includes an expiration date now so that the
token stays in the browser for longer. This is so that the user does
not have to reauthenticate when they restart their browser and the
previous access token has not expired yet.