Check that cookie values are not empty
The access token cookie must have a value so that it can be used to data
requests. The refresh token cookie must have a value to refresh the
access token.

If the access token is not empty and has not expired, it can still be
used.

If the refresh token is empty, it will print a message that it is
missing. Later on, it should prompt the user to reauthenticate to get a
new refresh token.