Increase search scope for Genius API, add popup
Sometimes, the Genius API responds with the correct Genius page when
only the first artist's name from Spotify is used in the query. However,
there are other times where the other artists' names must be included as
well because the track may have features or it is a remix. If they are
not included, the correct Genius page(s) will not show up.

This means that there are now two functions for searching. The fetch
statement that searches with the Genius API has been split into its
function called search(query) where it returns a promise which depends on its
response. This function is used in the other new function called
searchAll(query1, query2).